# Cache labs on server