--- a/.env.example
+++ b/.env.example
@@ -1,6 +1,7 @@
 DATABASE_URL = "postgres://root:mysecretpassword@localhost:5432/local"
 
 ABSOLUTE_DIR_PATH = ""
+LABS_CACHE_TTL_MIN = "10"
 
 BYPASS_ACCOUNT_REQUIREMENT = false
 CREATE_ACCOUNT = "example"
@@ -11,6 +12,6 @@ OIDC_CLIENT_SECRET = "e"
 OIDC_USERNAME_CLAIM = "preferred_username"
 
 ADDRESS_HEADER = "x-forwarded-for"
-XFF_DEPTH: = "1"
+XFF_DEPTH = "1"
 
 PUBLIC_UMAMI_ID = "e"