A large part of the page load time is currently reading and parsing the 2,200 line labs CSV file on every request. Because this file does not often change, it can be cached to drastically speed up many requests.

Environment variable `LABS_CACHE_TTL_MIN` controls how many minutes a file read is good for (suggested default 10).

GET /buildings 92ms -> 10ms
GET /labs 150ms -> 100ms
GET /summary 200ms -> 100ms

I think a lot of the load time for the lab-dependent routes is sending the 2,200 length array to the client - not much to do about that for now.